feat: 更新游戏评价相关功能及数据库结构
- 修改游戏评价表的评分字段类型，从decimal(2,1)更改为tinyint，支持1-5星评分。
- 更新GameReviewController中的接口描述，将创建评价的接口重命名为添加评价，并调整相关参数注释。
- 优化GameReviewServiceImpl中的评分验证逻辑，确保评分为1-5的整数。
- 更新相关请求类和视图对象，确保评分字段一致性。
- 调整SQL映射文件，优化字段格式。

+++ src/main/java/com/xm/game9/model/domain/GameReview.java
@@ -1,10 +1,6 @@
 package com.xm.game9.model.domain;
 
-import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.TableLogic;
-import com.baomidou.mybatisplus.annotation.TableName;
+import com.baomidou.mybatisplus.annotation.*;
 import lombok.Data;
 
 import java.io.Serial;
@@ -23,43 +19,43 @@ public class GameReview implements Serializable {
     @Serial
     @TableField(exist = false)
     private static final long serialVersionUID = 1L;
-    
+
     /**
      * 评价ID
      */
     @TableId(type = IdType.AUTO)
     private Long reviewId;
-    
+
     /**
      * 用户ID
      */
     private Long userId;
-    
+
     /**
      * 游戏ID
      */
     private Long gameId;
-    
+
     /**
-     * 评分（1-5分，支持0.5分）
+     * 评分（1-5星）
      */
-    private BigDecimal rating;
-    
+    private Integer rating;
+
     /**
      * 评价内容
      */
     private String content;
-    
+
     /**
      * 创建时间
      */
     private Date createTime;
-    
+
     /**
      * 更新时间
      */
     private Date updateTime;
-    
+
     /**
      * 是否删除
      */
